refactor: AnswerEntrySet の設定型を answer/settings から answer_entry_set/models へ統合する
DefaultAnswerTitle / AnswerVisibility / ResponsePeriod は AnswerEntrySet 集約のフィールドであり、
answer/settings に置くと個々の回答に紐づく設定に見えて誤解を招くため、
集約モデルと同じファイルへ移動して定義の分散を解消した。
