+++ server/domain/src/form/answer.rs
@@ -1,2 +1,1 @@
 pub mod models;
-pub mod settings;
