--- server/domain/src/form/answer/settings.rs
@@ -1,1 +1,0 @@
-pub mod models;
