+++ server/domain/src/form/models.rs
@@ -426,7 +426,7 @@ impl AuthorizationGuardDefinitions for ActiveForm {
     ///     user::models::{ActiveUser, Actor, Role, User},
     /// };
     /// use uuid::Uuid;
-    /// use domain::form::answer::settings::models::{AnswerVisibility, DefaultAnswerTitle, ResponsePeriod};
+    /// use domain::form::answer_entry_set::models::{AnswerVisibility, DefaultAnswerTitle, ResponsePeriod};
     /// use domain::form::answer_entry_set::models::AnswerEntrySetId;
     /// use domain::form::models::{
     ///     FormDescription, FormId, FormMeta,
