+++ server/domain/src/form/service.rs
@@ -4,10 +4,8 @@ use std::collections::HashMap;
 use std::sync::OnceLock;
 
 use crate::form::{
-    answer::{
-        models::{AnswerTitle, PostedAnswerContents},
-        settings::models::DefaultAnswerTitle,
-    },
+    answer::models::{AnswerTitle, PostedAnswerContents},
+    answer_entry_set::models::DefaultAnswerTitle,
     models::Question,
 };
